refactor(core): replace magic numbers with named constants
Extract hardcoded numeric literals into named constants to improve
code readability and maintainability. This refactoring ensures that
domain-specific values are defined once and referenced consistently.

Changes include:

- Add ULID_TIMESTAMP_CHARS and ULID_RANDOMNESS_CHARS constants
- Replace hardcoded 1024 with MAX_HASH_OUTPUT_BYTES in error messages
- Replace hardcoded 10000 with MAX_BULK_GENERATION in validation
- Replace hardcoded 26 with ULID_STRING_LENGTH throughout tests
- Replace hardcoded 32 with CROCKFORD_BASE32_CHARSET.len()
- Update error messages to use format strings with constants

Also extracts UlidInspectCommand into its own module (inspect.rs)
with dedicated constants for ULID bit widths and time calculations:
- ULID_TIMESTAMP_BITS, ULID_RANDOMNESS_BITS, ULID_TOTAL_BITS
- SECONDS_PER_MINUTE, SECONDS_PER_HOUR, SECONDS_PER_DAY

This improves code maintainability by centralizing magic numbers
and makes future value changes safer and more consistent.
